+++ src/dfx/src/lib/builders/motoko.rs
@@ -166,7 +166,8 @@ impl CanisterBuilder for MotokoBuilder {
         };
         motoko_compile(&self.logger, cache.as_ref(), &params)?;
 
-        if canister_info.get_shrink() {
+        let shrink = canister_info.get_shrink().unwrap_or(true);
+        if shrink {
             info!(self.logger, "Shrink WASM module size.");
             super::shrink_wasm(motoko_info.get_output_wasm_path())?;
         }